Swarmos/requirements.txt
aiohttp
psutil
cryptography
